tests/submission/Microsoft/forms/autofocus.html

<!DOCTYPE HTML>
<html>
    <head>
        <title>autofocus on input element</title>
        <meta http-equiv="content-type" content="text/html; charset=UTF-8" />
        <meta name="description" content="autofocus on input element" />
    </head>
    <body>
	<script src="/resources/testharness.js"></script>
    <script src="http://w3c-test.org/resources/testharnessreport.js"></script>
    <div id=log></div>

    <input style="display: none" type="text" autofocus></input>
    <input style="display: none" type="text" autofocus="false"></input>
    <input style="display: none" type="text" ></input>
    
    <script type="text/javascript">
        test(function() {assert_true(document.getElementsByTagName("input")[0].autofocus)}, "autofocus on input element on input element");
        test(function() {assert_true(document.getElementsByTagName("input")[1].autofocus)}, "autofocus returns false on input element");
        test(function() {assert_equals(document.getElementsByTagName("input")[1].getAttribute("autofocus"), "false")}, "autofocus returns false when retrieved using getAttribute");
        
        test(function() {assert_false(document.getElementsByTagName("input")[2].autofocus)}, "input element without autofocus attribute returns false");
        test(function() {assert_equals(document.getElementsByTagName("input")[2].getAttribute("autofocus"), null)}, "input element's autofocus attribute is null when not specified");
        
        document.getElementsByTagName("input")[2].autofocus = true;
        
        test(function() {assert_true(document.getElementsByTagName("input")[2].autofocus)}, "autofocus on input element is true when set via script");
        test(function() {assert_equals(document.getElementsByTagName("input")[2].getAttribute("autofocus"), "")}, "accessing an input element's autofocus attribute is an empty string");
        
    </script>
    </body>
</html>
